Fixed some issues
I fixed a few typos, as well as moved the "print(t)" function outside of the loop but within the if/else statement. Before it would stop after one iteration. Now it runs smoothly.

I ran a few random combinations of temperature, and this function is getting similar results to the "old" function.