Moved calcQ function
I was having issues with the "calcQ" function running with a new Ta value placed within any of the functions that had "calcQ" within it's component list. I believe it has something to do with R not recognizing a new Ta value. Therefore I moved "calcQ" to be within the {} body of this function, so it would be calculated with whatever Ta value is being used.